pyproject.toml
[tool.poetry]
name = "pyconntrack"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
wemake-python-styleguide = "^0.17.0"
flake8-pytest-style = "^1.6.0"
mypy = "^0.991"
safety = "^2.3.5"
pre-commit = "^2.20.0"
requests-cache = "^0.9.7"
nitpick = "^0.32.0"
netifaces = "^0.11.0"
ipy = "^1.1"
rich = "^12.6.0"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"
